ms7/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
fpdf2==2.7.9
boto3==1.35.0
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
